Ostafandy/requirements.txt
Django==2.2.6
djangorestframework==3.9.2
django-cors-headers==3.2.0
psycopg2-binary==2.8.3
gunicorn==19.9.0



